:recycle: Refactor #42: 버블 보관함 & 라벨 상세 화면 하나로 합치기
