+++ app/src/main/java/com/umc/edison/ui/bubblestorage/BubbleStorageScreen.kt
@@ -5,13 +5,12 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.*
-import androidx.compose.material3.Divider
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
-import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -19,7 +18,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
@@ -27,14 +25,15 @@ import com.umc.edison.ui.components.Bubble
 import com.umc.edison.presentation.model.BubbleModel
 import com.umc.edison.presentation.storage.BubbleStorageMode
 import com.umc.edison.presentation.storage.BubbleStorageViewModel
-import com.umc.edison.ui.LoadingScreen
+import com.umc.edison.ui.BaseContent
 import com.umc.edison.ui.components.BottomSheet
 import com.umc.edison.ui.components.BottomSheetForDelete
 import com.umc.edison.ui.components.BottomSheetPopUp
 import com.umc.edison.ui.components.BubblesLayout
+import com.umc.edison.ui.components.LabelTopAppBar
+import com.umc.edison.ui.label.LabelMoveModalContent
 import com.umc.edison.ui.theme.Gray300
 import com.umc.edison.ui.theme.Gray900
-import com.umc.edison.ui.theme.White000
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -43,11 +42,8 @@ fun BubbleStorageScreen(
     updateShowBottomNav: (Boolean) -> Unit,
     viewModel: BubbleStorageViewModel = hiltViewModel(),
 ) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-
     val uiState by viewModel.uiState.collectAsState()
-    val isBlur =
-        uiState.bubbleStorageMode != BubbleStorageMode.NONE && uiState.bubbleStorageMode != BubbleStorageMode.VIEW
+    val isBlur = uiState.bubbleStorageMode != BubbleStorageMode.NONE
 
     BackHandler(enabled = true) {
         if (uiState.bubbleStorageMode == BubbleStorageMode.NONE) {
@@ -58,179 +54,194 @@ fun BubbleStorageScreen(
     }
 
     Scaffold(
-
         bottomBar = {
             if (uiState.bubbleStorageMode == BubbleStorageMode.EDIT) {
+                val onButtonClick: () -> Unit
+                val buttonText: String
+
+                if (uiState.label == null) {
+                    onButtonClick = {
+                        viewModel.updateEditMode(BubbleStorageMode.SHARE)
+                    }
+                    buttonText = "공유하기"
+                } else {
+                    onButtonClick = {
+                        viewModel.getMovableLabels()
+                        viewModel.updateEditMode(BubbleStorageMode.MOVE)
+                    }
+                    buttonText = "버블 이동"
+                }
+
                 BottomSheetForDelete(
                     selectedCnt = uiState.selectedBubbles.size,
                     showSelectedCnt = true,
-                    onButtonClick = {
-                        // 공유하기 버튼 눌렀을 때 동작
-                        viewModel.updateEditMode(BubbleStorageMode.SHARE)
-                    },
+                    onButtonClick = onButtonClick,
                     onDelete = {
                         viewModel.updateEditMode(BubbleStorageMode.DELETE)
                     },
                     buttonEnabled = uiState.selectedBubbles.isNotEmpty(),
-                    buttonText = "공유하기",
+                    buttonText = buttonText,
                 )
             }
         }
     ) { innerPadding ->
-        Box(
+        BaseContent(
+            uiState = uiState,
+            onDismiss = { viewModel.clearError() },
+            modifier = Modifier.padding(innerPadding),
+        ) {
+            var onBubbleClick: (BubbleModel) -> Unit = {}
+            var onBubbleLongClick: (BubbleModel) -> Unit = {}
 
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(innerPadding)
-                .background(White000)
-        )
-        {
-            if (uiState.isLoading) {
-                LoadingScreen()
-            } else if (uiState.error != null) {
-                Text(
-                    text = "Error loading data",
-                    color = Color.Red,
-                    modifier = Modifier.padding(16.dp),
-                    textAlign = TextAlign.Center
-                )
-            } else {
-                var onBubbleClick: (BubbleModel) -> Unit = {}
-                var onBubbleLongClick: (BubbleModel) -> Unit = {}
+            if (uiState.bubbleStorageMode == BubbleStorageMode.EDIT) {
+                onBubbleClick = { bubble ->
+                    viewModel.toggleSelectBubble(bubble)
+                }
+            } else if (uiState.bubbleStorageMode == BubbleStorageMode.NONE) {
+                onBubbleClick = { bubble ->
+                    viewModel.selectBubble(bubble)
+                    viewModel.updateEditMode(BubbleStorageMode.VIEW)
+                }
+                onBubbleLongClick = { bubble ->
+                    viewModel.selectBubble(bubble)
+                    viewModel.updateEditMode(BubbleStorageMode.EDIT)
+                    updateShowBottomNav(false)
+                }
+            }
 
-                if (uiState.bubbleStorageMode == BubbleStorageMode.EDIT) {
-                    onBubbleClick = { bubble ->
-                        viewModel.toggleSelectBubble(bubble)
-                    }
-                } else if (uiState.bubbleStorageMode == BubbleStorageMode.NONE) {
-                    onBubbleClick = { bubble ->
-                        viewModel.selectBubble(bubble)
-                        viewModel.updateEditMode(BubbleStorageMode.VIEW)
-                    }
-                    onBubbleLongClick = { bubble ->
-                        viewModel.selectBubble(bubble)
-                        viewModel.updateEditMode(BubbleStorageMode.EDIT)
-                        updateShowBottomNav(false)
+            Column(
+                modifier = Modifier.clickable(
+                    indication = null,
+                    interactionSource = remember { MutableInteractionSource() }
+                ) {
+                    if (uiState.bubbleStorageMode == BubbleStorageMode.EDIT) {
+                        resetEditMode(viewModel, updateShowBottomNav)
                     }
                 }
+            ) {
+                if (uiState.label != null) {
+                    LabelTopAppBar(
+                        label = uiState.label!!,
+                        onBackClick = {
+                            resetEditMode(viewModel, updateShowBottomNav)
+                            navHostController.popBackStack()}
+                    )
+                }
 
+                BubblesLayout(
+                    bubbles = uiState.label?.bubbles ?: uiState.bubbles,
+                    onBubbleClick = onBubbleClick,
+                    onBubbleLongClick = onBubbleLongClick,
+                    isBlur = isBlur,
+                    selectedBubble = uiState.selectedBubbles,
+                )
+            }
+
+            if (uiState.bubbleStorageMode == BubbleStorageMode.VIEW && uiState.selectedBubbles.isNotEmpty()) {
+                val bubble = uiState.selectedBubbles.first()
                 Box(
                     modifier = Modifier
-                        .clickable(
-                            indication = null,
-                            interactionSource = remember { MutableInteractionSource() }
-                        ) {
-                            if (uiState.bubbleStorageMode == BubbleStorageMode.EDIT) {
-                                resetEditMode(viewModel, updateShowBottomNav)
-                            }
-                        }
                         .fillMaxSize()
+                        .background(Color.Black.copy(alpha = 0.5f))
+                        .clickable(onClick = {
+                            viewModel.updateEditMode(BubbleStorageMode.NONE)
+                        }),
+                    contentAlignment = Alignment.Center
                 ) {
-                    BubblesLayout(
-                        bubbles = uiState.bubbles,
-                        onBubbleClick = onBubbleClick,
-                        onBubbleLongClick = onBubbleLongClick,
-                        isBlur = isBlur,
-                        selectedBubble = uiState.selectedBubbles,
+                    Bubble(
+                        bubble = bubble,
+                        onClick = {
+                            // TODO: 버블 작성 화면 구현 완료되면 연결
+                            // navHostController.navigate(NavRoute.BubbleEdit.createRoute(bubble.id))
+                        }
                     )
                 }
-
-                if (uiState.bubbleStorageMode == BubbleStorageMode.VIEW && uiState.selectedBubbles.isNotEmpty()) {
-                    val bubble = uiState.selectedBubbles.first()
-                    Box(
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .background(Color.Black.copy(alpha = 0.5f))
-                            .clickable(onClick = {
-                                viewModel.updateEditMode(BubbleStorageMode.NONE)
-                            }),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Bubble(
-                            bubble = bubble,
-                            onClick = {
-                                // TODO: 버블 작성 화면 구현 완료되면 연결
-                                // navHostController.navigate(NavRoute.BubbleEdit.createRoute(bubble.id))
-                            }
-                        )
-                    }
-                } else if (uiState.bubbleStorageMode == BubbleStorageMode.SHARE) {
-                    BottomSheet(
+            } else if (uiState.bubbleStorageMode == BubbleStorageMode.MOVE) {
+                BottomSheet(
+                    onDismiss = {
+                        viewModel.updateEditMode(BubbleStorageMode.EDIT)
+                    },
+                ) {
+                    LabelMoveModalContent(
+                        labels = uiState.movableLabels,
                         onDismiss = {
                             viewModel.updateEditMode(BubbleStorageMode.EDIT)
                         },
-                        sheetState = sheetState,
+                        onConfirm = { label ->
+                            viewModel.moveSelectedBubbles(label, showBottomNav = updateShowBottomNav)
+                        }
+                    )
+                }
+            } else if (uiState.bubbleStorageMode == BubbleStorageMode.DELETE) {
+                BottomSheetPopUp(
+                    title = "${uiState.selectedBubbles.size} 개의 버블을 삭제하시겠습니까?",
+                    cancelText = "취소",
+                    confirmText = "삭제",
+                    onDismiss = {
+                        viewModel.updateEditMode(BubbleStorageMode.EDIT)
+                    },
+                    onConfirm = {
+                        viewModel.deleteSelectedBubbles(showBottomNav = updateShowBottomNav)
+                    },
+                )
+            } else if (uiState.bubbleStorageMode == BubbleStorageMode.SHARE) {
+                BottomSheet(
+                    onDismiss = {
+                        viewModel.updateEditMode(BubbleStorageMode.EDIT)
+                    },
+                ) {
+                    Column(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(vertical = 16.dp)
                     ) {
-                        Column(
+                        TextButton(
+                            onClick = { /* TODO: 이미지 공유 로직 추가 */ },
                             modifier = Modifier
                                 .fillMaxWidth()
-                                .padding(vertical = 16.dp)
+                                .padding(horizontal = 16.dp, vertical = 10.dp)
                         ) {
-                            TextButton(
-                                onClick = { /* TODO: 이미지 공유 로직 추가 */ },
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .padding(horizontal = 16.dp, vertical = 10.dp)
+                            Row(
+                                modifier = Modifier.fillMaxWidth()
                             ) {
-                                Row(
-                                    modifier = Modifier.fillMaxWidth()
-                                ) {
-                                    Text(
-                                        text = "이미지로 공유하기",
-                                        style = MaterialTheme.typography.titleLarge,
-                                        color = Gray900
-                                    )
-                                }
-                            }
-
-                            Box(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .wrapContentWidth(Alignment.CenterHorizontally)
-                            ) {
-                                Divider(
-                                    color = Gray300,
-                                    thickness = 1.dp,
-                                    modifier = Modifier.width(326.dp)
+                                Text(
+                                    text = "이미지로 공유하기",
+                                    style = MaterialTheme.typography.titleLarge,
+                                    color = Gray900
                                 )
-                            }
-
-                            TextButton(
-                                onClick = { /* TODO: 텍스트 공유 로직 추가 */ },
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .padding(start = 16.dp, end = 16.dp, top = 16.dp, bottom = 10.dp)
-
-                            ) {
-                                Row(
-                                    modifier = Modifier.fillMaxWidth()
-                                ) {
-                                    Text(
-                                        text = "텍스트로 공유하기",
-                                        style = MaterialTheme.typography.titleLarge,
-                                        color = Gray900
-                                    )
-                                }
                             }
                         }
 
+                        Box(
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .wrapContentWidth(Alignment.CenterHorizontally)
+                        ) {
+                            HorizontalDivider(
+                                color = Gray300,
+                                thickness = 1.dp,
+                                modifier = Modifier.width(326.dp)
+                            )
+                        }
 
-                }
+                        TextButton(
+                            onClick = { /* TODO: 텍스트 공유 로직 추가 */ },
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .padding(start = 16.dp, end = 16.dp, top = 16.dp, bottom = 10.dp)
 
-
-                } else if (uiState.bubbleStorageMode == BubbleStorageMode.DELETE) {
-                    BottomSheetPopUp(
-                        title = "${uiState.selectedBubbles.size} 개의 버블을 삭제하시겠습니까?",
-                        cancelText = "취소",
-                        confirmText = "삭제",
-                        onDismiss = {
-                            viewModel.updateEditMode(BubbleStorageMode.EDIT)
-                        },
-                        onConfirm = {
-                             viewModel.deleteSelectedBubbles(showBottomNav = updateShowBottomNav)
-                        },
-                        sheetState = sheetState,
-                    )
+                        ) {
+                            Row(
+                                modifier = Modifier.fillMaxWidth()
+                            ) {
+                                Text(
+                                    text = "텍스트로 공유하기",
+                                    style = MaterialTheme.typography.titleLarge,
+                                    color = Gray900
+                                )
+                            }
+                        }
+                    }
                 }
             }
         }
