+++ app/src/main/java/com/umc/edison/ui/components/TopBar.kt
@@ -1,13 +1,19 @@
 package com.umc.edison.ui.components
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -20,6 +26,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import com.umc.edison.R
+import com.umc.edison.presentation.model.LabelModel
+import com.umc.edison.ui.theme.Gray100
 import com.umc.edison.ui.theme.Gray800
 import com.umc.edison.ui.theme.White000
 
@@ -95,3 +103,33 @@ fun BackButtonTopBar(
         )
     }
 }
+
+@Composable
+fun LabelTopAppBar(
+    label: LabelModel,
+    onBackClick: () -> Unit
+) {
+    BackButtonTopBar(
+        onBack = onBackClick,
+    ) {
+        Box(
+            modifier = Modifier
+                .size(24.dp)
+                .background(color = label.color, shape = CircleShape)
+                .border(
+                    width = 3.dp,
+                    color = if (label.color != White000) label.color else Gray100,
+                    shape = RoundedCornerShape(16.dp)
+                )
+        )
+
+        Spacer(modifier = Modifier.width(12.dp))
+
+        Text(
+            text = "${label.name}  ${label.bubbles.size}",
+            style = MaterialTheme.typography.titleLarge,
+            color = Gray800
+        )
+    }
+}
+
